reworking characteric value views
